rename Sample to Trace
